+++ R/tidyverse.R
@@ -160,7 +160,7 @@ use_tidy_issue_template <- function() {
 #' @rdname tidyverse
 use_tidy_coc <- function() {
   use_dot_github()
-  use_code_of_conduct(path = ".github")
+  use_code_of_conduct("codeofconduct@rstudio.com", path = ".github")
 }
 
 #' @export